refactor: move SearchRepositoryByName testing.
