+++ models/repo_test.go
@@ -125,55 +125,3 @@ func TestForkRepository(t *testing.T) {
 	assert.Error(t, err)
 	assert.True(t, IsErrRepoAlreadyExist(err))
 }
-
-func TestSearchRepositoryByName(t *testing.T) {
-	assert.NoError(t, PrepareTestDatabase())
-
-	// test search public repository on explore page
-	repos, count, err := SearchRepositoryByName(&SearchRepoOptions{
-		Keyword:  "repo_12",
-		Page:     1,
-		PageSize: 10,
-		Searcher: nil,
-	})
-
-	assert.NotNil(t, repos)
-	assert.NoError(t, err)
-	assert.Equal(t, "test_repo_12", repos[0].Name)
-	assert.Equal(t, int64(1), count)
-
-	repos, count, err = SearchRepositoryByName(&SearchRepoOptions{
-		Keyword:  "test_repo",
-		Page:     1,
-		PageSize: 10,
-		Searcher: nil,
-	})
-
-	assert.NotNil(t, repos)
-	assert.NoError(t, err)
-	assert.Equal(t, int64(2), count)
-
-	// test search private repository on explore page
-	repos, count, err = SearchRepositoryByName(&SearchRepoOptions{
-		Keyword:  "repo_13",
-		Page:     1,
-		PageSize: 10,
-		Searcher: &User{ID: 14},
-	})
-
-	assert.NotNil(t, repos)
-	assert.NoError(t, err)
-	assert.Equal(t, "test_repo_13", repos[0].Name)
-	assert.Equal(t, int64(1), count)
-
-	repos, count, err = SearchRepositoryByName(&SearchRepoOptions{
-		Keyword:  "test_repo",
-		Page:     1,
-		PageSize: 10,
-		Searcher: &User{ID: 14},
-	})
-
-	assert.NotNil(t, repos)
-	assert.NoError(t, err)
-	assert.Equal(t, int64(3), count)
-}
